+++ client/app/[locale]/admin/dashboard/heat-map/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Page() {
+  return (
+    <div>Heat Map</div>
+  )
+}
