+++ client/app/[locale]/admin/dashboard/layout.tsx
@@ -1,0 +1,15 @@
+'use client'
+import TabMenu from "@/components/tab-menu"
+
+export default function Layout({
+    children,
+  }: {
+    children: React.ReactNode
+  }) {
+    return (
+      <section className="flex flex-col px-6 py-4 gap-4">
+        <TabMenu />
+        {children}
+      </section>
+    )
+  }
