+++ client/app/[locale]/admin/dashboard/overview/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Page() {
+  return (
+    <div>Overview</div>
+  )
+}
