+++ client/app/[locale]/admin/settings/patrol-preset/page.tsx
@@ -1,0 +1,7 @@
+import React from 'react'
+
+export default function Page() {
+  return (
+    <div>page</div>
+  )
+}
